+++ src/pages/about/components/index.jsx
@@ -1,0 +1,9 @@
+import Header from './header';
+import Meeting from './meeting';
+import Exec from './exec';
+
+export {
+  Header,
+  Meeting,
+  Exec,
+};
