+++ src/pages/home/components/headerMessage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useViewport } from '../../components';
+import { useViewport } from '../../../components';
 
 function HeaderMessage() {
   const { width } = useViewport();
