+++ src/pages/home/components/index.jsx
@@ -1,0 +1,11 @@
+import Header from './header';
+import Subheader from './whoAreWe';
+import Subheader2 from './whatDoWeDo';
+import Subheader3 from './contactUs';
+
+export {
+  Header,
+  Subheader,
+  Subheader2,
+  Subheader3,
+};
