+++ src/pages/policy/components/index.jsx
@@ -1,0 +1,7 @@
+import FileViewer from './fileViewer';
+import Header from './header';
+
+export {
+  FileViewer,
+  Header,
+};
