+++ src/pages/policy/index.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import Container from '@material-ui/core/Container';
-import Header from './header';
-import FileViewer from './fileViewer';
+import { FileViewer, Header } from './components';
 
-import constitution from '../../assets/Constitution.pdf';
-import clubSpacePolicy from '../../assets/ClubSpacePolicy.pdf';
-import committeeAndStaffPolicy from '../../assets/CommitteeAndStaffPolicy.pdf';
-import makersSpacePolicy from '../../assets/MakersSpacePolicy.pdf';
+import constitution from './assets/Constitution.pdf';
+import clubSpacePolicy from './assets/ClubSpacePolicy.pdf';
+import committeeAndStaffPolicy from './assets/CommitteeAndStaffPolicy.pdf';
+import makersSpacePolicy from './assets/MakersSpacePolicy.pdf';
 
 function PolicyPage() {
   return (
